[Client] Refactoring modal name and delete remarks
